Add database

+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -6,9 +6,13 @@ import java.util.List;
 
 public interface FilmStorage {
     void create(Film film);
+
     void update(Film film);
-    void delete(long id);
+
+    void delete(int id);
+
     List<Film> getAll();
-    Film getById(long id);
+
+    Film getById(int id);
 
 }
